# Generate unique ids for headers with duplicate text