--- a/lib/marked.js
+++ b/lib/marked.js
@@ -148,6 +148,7 @@ Lexer.prototype.lex = function(src) {
 
 Lexer.prototype.token = function(src, top, bq) {
   var src = src.replace(/^ +$/gm, '')
+    , links = {}
     , next
     , loose
     , cap
@@ -157,6 +158,24 @@ Lexer.prototype.token = function(src, top, bq) {
     , space
     , i
     , l;
+
+  function uniqueAnchor(text){
+    var rawAnchor = text.toLowerCase().replace(/[^\w]+/g, '-');
+    var anchor = rawAnchor;
+
+    if(!links[anchor]){
+        links[anchor] = anchor;
+        return anchor;
+    } else {
+      var num = 0;
+      while (links[anchor]){
+        num++;
+        anchor = rawAnchor + ((num) ? '-'+num : '')
+      }
+      links[anchor] = anchor;
+      return anchor;
+    }
+  }
 
   while (src) {
     // newline
@@ -199,7 +218,8 @@ Lexer.prototype.token = function(src, top, bq) {
       this.tokens.push({
         type: 'heading',
         depth: cap[1].length,
-        text: cap[2]
+        text: cap[2],
+        anchor: uniqueAnchor(cap[2])
       });
       continue;
     }
@@ -242,7 +262,8 @@ Lexer.prototype.token = function(src, top, bq) {
       this.tokens.push({
         type: 'heading',
         depth: cap[2] === '=' ? 1 : 2,
-        text: cap[1]
+        text: cap[1],
+        anchor: uniqueAnchor(cap[1])
       });
       continue;
     }
@@ -785,12 +806,12 @@ Renderer.prototype.html = function(html) {
   return html;
 };
 
-Renderer.prototype.heading = function(text, level, raw) {
+Renderer.prototype.heading = function(text, level, raw, anchor) {
   return '<h'
     + level
     + ' id="'
     + this.options.headerPrefix
-    + raw.toLowerCase().replace(/[^\w]+/g, '-')
+    + anchor
     + '">'
     + text
     + '</h'
@@ -973,7 +994,8 @@ Parser.prototype.tok = function() {
       return this.renderer.heading(
         this.inline.output(this.token.text),
         this.token.depth,
-        this.token.text);
+        this.token.text,
+        this.token.anchor);
     }
     case 'code': {
       return this.renderer.code(this.token.text,