--- a/test/tests/duplicate_headers.html
+++ b/test/tests/duplicate_headers.html
@@ -1,0 +1,7 @@
+<h1 id="header">Header</h1>
+
+<h2 id="example">Example</h2>
+
+<h1 id="header-1">Header</h1>
+
+<h2 id="example-1">Example</h2>