--- a/test/tests/duplicate_headers.text
+++ b/test/tests/duplicate_headers.text
@@ -1,0 +1,7 @@
+# Header
+
+## Example
+
+# Header
+
+## Example