@chjj 2015-05-08
That 200ms is devestating. There may be a better way to optimize this. I think this uniqueness issue was discussed a while back, but we decided to leave things as they are for the time being. Can we work on optimizing this more?