@josephernest 2016-12-03
@chjj @patrickarlt Is there some news on this?

I also have sometimes identical `<h2>` names and then the anchor links / id are not unique, that's a problem.
```

# Thing 1

## Examples

## Options

# Thing 2

## Examples
```